using sstreams to match the condition of lab problem

+++ reader.h
@@ -5,7 +5,7 @@
 class Reader {
 public:
     struct InputData {
-        size_t n;
+        long long n;
         std::vector<int> values;
         std::vector<double> weights;
     };
